requirements.txt
python-socketio[client]
tweepy